Move mainwindow.py to mainwindow/__init__.py, and mainwindow_actions to mainwindow/actions.py
